Calculate cursor radius for color picker plane based on slider width
Slider are able to calculate their own cursor radius because cursor
radius for the component is defined as r = sw/2 where sw is the width of
a slider, already passed down through css and the DOM. The color plane
does not have this advantage, giving the options of either using the
extremely sluggish getComputedStyles method, or manually passing cursor
radius as a parameter. The end result is that all cursors in the
component are the exact same size.