Add basic TOTP logic
Not sure how much of this logic will be functional once I have TOTP setup to a point where it can be tested. This still provides a good starting point, and the refactoring of how user credentials are passed to actions.authenticate will save a lot of future headaches as 2FA complexity grows.