Rewrite lists store
Lists store now maps to API structure 1:1 (metadata is kept in the meta field), I'm not sure if I want to keep this behavior or go back to moving everything to the top level for client storage. This was also an opportunity to fix a nasty bug where list indexes weren't being properly updated, resulting in a multitude of broken behavior. The API will be including indexes in checksum calculations to prevent this or similar bugs from happening in the future.