Remove all private key export functionality
Upon mentally weighing the pros and cons of allowing private keys to be
exported, I've decided that there is no reason for implementing this as
user-facing behavior. Public keys may wish to be shared for the purpose
of key fingerprint verification through an external channel, but private keys have no reason to be exported by normal users.